Integrate Snapshot types and interfaces into rest of codebase
