+++ src/utils/snapshotVote.ts
@@ -1,4 +1,16 @@
 import { Account, Address, Chain, Transport, WalletClient } from 'viem';
+import {
+  SnapshotVote,
+  snapshotVote2Types,
+  snapshotVoteArray2Types,
+  snapshotVoteArrayTypes,
+  SnapshotVoteForSigning,
+  snapshotVoteStringTypes,
+  snapshotVoteString2Types,
+  snapshotVoteTypes,
+  SnapshotVoteTypes,
+  SnapshotSignMessage,
+} from '../types/snapshot';
 
 const NAME = 'snapshot';
 const VERSION = '0.1.4';
@@ -9,170 +21,38 @@ const domain = {
   version: VERSION,
 } as const;
 
-type ProposalType =
-  | 'single-choice'
-  | 'approval'
-  | 'quadratic'
-  | 'ranked-choice'
-  | 'weighted'
-  | 'basic';
-
-const voteTypes = {
-  Vote: [
-    { name: 'from', type: 'address' },
-    { name: 'space', type: 'string' },
-    { name: 'timestamp', type: 'uint64' },
-    { name: 'proposal', type: 'string' },
-    { name: 'choice', type: 'uint32' },
-    { name: 'reason', type: 'string' },
-    { name: 'app', type: 'string' },
-    { name: 'metadata', type: 'string' },
-  ],
-} as const;
-
-const vote2Types = {
-  Vote: [
-    { name: 'from', type: 'address' },
-    { name: 'space', type: 'string' },
-    { name: 'timestamp', type: 'uint64' },
-    { name: 'proposal', type: 'bytes32' },
-    { name: 'choice', type: 'uint32' },
-    { name: 'reason', type: 'string' },
-    { name: 'app', type: 'string' },
-    { name: 'metadata', type: 'string' },
-  ],
-} as const;
-
-const voteArray2Types = {
-  Vote: [
-    { name: 'from', type: 'address' },
-    { name: 'space', type: 'string' },
-    { name: 'timestamp', type: 'uint64' },
-    { name: 'proposal', type: 'bytes32' },
-    { name: 'choice', type: 'uint32[]' },
-    { name: 'reason', type: 'string' },
-    { name: 'app', type: 'string' },
-    { name: 'metadata', type: 'string' },
-  ],
-} as const;
-
-const voteArrayTypes = {
-  Vote: [
-    { name: 'from', type: 'address' },
-    { name: 'space', type: 'string' },
-    { name: 'timestamp', type: 'uint64' },
-    { name: 'proposal', type: 'string' },
-    { name: 'choice', type: 'uint32[]' },
-    { name: 'reason', type: 'string' },
-    { name: 'app', type: 'string' },
-    { name: 'metadata', type: 'string' },
-  ],
-} as const;
-
-const voteString2Types = {
-  Vote: [
-    { name: 'from', type: 'address' },
-    { name: 'space', type: 'string' },
-    { name: 'timestamp', type: 'uint64' },
-    { name: 'proposal', type: 'bytes32' },
-    { name: 'choice', type: 'string' },
-    { name: 'reason', type: 'string' },
-    { name: 'app', type: 'string' },
-    { name: 'metadata', type: 'string' },
-  ],
-} as const;
-
-const voteStringTypes = {
-  Vote: [
-    { name: 'from', type: 'address' },
-    { name: 'space', type: 'string' },
-    { name: 'timestamp', type: 'uint64' },
-    { name: 'proposal', type: 'string' },
-    { name: 'choice', type: 'string' },
-    { name: 'reason', type: 'string' },
-    { name: 'app', type: 'string' },
-    { name: 'metadata', type: 'string' },
-  ],
-} as const;
-
-// Now, let's create a union type for all possible vote types
-type VoteTypes =
-  | typeof voteTypes
-  | typeof vote2Types
-  | typeof voteArrayTypes
-  | typeof voteArray2Types
-  | typeof voteStringTypes
-  | typeof voteString2Types;
-
-interface Vote {
-  from: Address;
-  space: string;
-  proposal: string;
-  type: ProposalType;
-  choice:
-    | number
-    | number[]
-    | string
-    | {
-        [key: string]: number;
-      };
-  privacy?: string;
-  reason?: string;
-  app?: string;
-  metadata?: string;
-}
-
-interface VoteForSigning extends Omit<Vote, 'type' | 'privacy'> {
-  timestamp: bigint;
-  reason: string;
-  app: string;
-  metadata: string;
-}
-
-// Add these type definitions at the top of the file or in a separate types file
-type SignMessageBase = Omit<VoteForSigning, 'choice'>;
-
-type SignMessageSingleChoice = SignMessageBase & {
-  choice: number;
-};
-
-type SignMessageArrayChoice = SignMessageBase & {
-  choice: number[];
-};
-
-type SignMessageStringChoice = SignMessageBase & {
-  choice: string;
-};
-
-type SignMessage = SignMessageSingleChoice | SignMessageArrayChoice | SignMessageStringChoice;
-
 const customSerializer = (_: string, value: any) => {
   if (typeof value === 'bigint') {
-    return value.toString();
+    // The Snapshot API expects the timestamp to be a number,
+    // even though the type is uint64 (which is a bigint).
+    // So we must just do this unsafe conversion. Oh well.
+    return Number(value);
   }
   return value;
 };
 
-const parseMessage = (message: Vote): { message: VoteForSigning; types: VoteTypes } => {
+const parseMessage = (
+  message: SnapshotVote,
+): { message: SnapshotVoteForSigning; types: SnapshotVoteTypes } => {
   const isShutter = message?.privacy === 'shutter';
   const type2 = message.proposal.startsWith('0x');
-  let types: VoteTypes = type2 ? vote2Types : voteTypes;
+  let types: SnapshotVoteTypes = type2 ? snapshotVote2Types : snapshotVoteTypes;
 
   if (['approval', 'ranked-choice'].includes(message.type)) {
-    types = type2 ? voteArray2Types : voteArrayTypes;
+    types = type2 ? snapshotVoteArray2Types : snapshotVoteArrayTypes;
   }
 
   let choice = message.choice;
   if (!isShutter && ['quadratic', 'weighted'].includes(message.type)) {
-    types = type2 ? voteString2Types : voteStringTypes;
+    types = type2 ? snapshotVoteString2Types : snapshotVoteStringTypes;
     choice = JSON.stringify(message.choice);
   }
 
   if (isShutter) {
-    types = type2 ? voteString2Types : voteStringTypes;
+    types = type2 ? snapshotVoteString2Types : snapshotVoteStringTypes;
   }
 
-  const messageForSigning: VoteForSigning = {
+  const messageForSigning: SnapshotVoteForSigning = {
     from: message.from || '',
     space: message.space,
     timestamp: BigInt(Math.floor(Date.now() / 1000)),
@@ -188,14 +68,14 @@ const parseMessage = (message: Vote): { message: VoteForSigning; types: VoteType
 
 const sign = async (
   walletClient: WalletClient<Transport, Chain, Account>,
-  message: VoteForSigning,
-  types: VoteTypes,
+  message: SnapshotVoteForSigning,
+  types: SnapshotVoteTypes,
 ): Promise<{ address: Address; sig: string; data: any }> => {
-  let signMessage: SignMessage;
+  let signMessage: SnapshotSignMessage;
 
-  if (types === voteTypes || types === vote2Types) {
+  if (types === snapshotVoteTypes || types === snapshotVote2Types) {
     signMessage = { ...message, choice: Number(message.choice) };
-  } else if (types === voteArrayTypes || types === voteArray2Types) {
+  } else if (types === snapshotVoteArrayTypes || types === snapshotVoteArray2Types) {
     signMessage = {
       ...message,
       choice: Array.isArray(message.choice) ? message.choice : [Number(message.choice)],
@@ -214,10 +94,9 @@ const sign = async (
     domain,
     types,
     message: signMessage,
-    primaryType: 'Vote' as const,
   };
 
-  const sig = await walletClient.signTypedData(data);
+  const sig = await walletClient.signTypedData({ ...data, primaryType: 'Vote' });
 
   return { address: message.from, sig, data };
 };
@@ -227,7 +106,7 @@ const send = async (
   envelop: {
     address: Address;
     sig: string;
-    data: { domain: typeof domain; types: VoteTypes; message: VoteForSigning };
+    data: { domain: typeof domain; types: SnapshotVoteTypes; message: SnapshotVoteForSigning };
   },
 ) => {
   const init = {
@@ -249,7 +128,7 @@ const send = async (
 
 export const submitSnapshotVote = async (
   walletClient: WalletClient<Transport, Chain, Account>,
-  message: Vote,
+  message: SnapshotVote,
 ) => {
   const { message: messageForSigning, types: typesForSigning } = parseMessage(message);
   const { address, sig, data } = await sign(walletClient, messageForSigning, typesForSigning);
